feat: leave room in footer
