+++ components/hms/footer/index.module.css
@@ -6,6 +6,7 @@
   justify-content: center;
   align-items: center;
   gap: 1.2rem;
+  position: relative;
 }
 .btn-wrapper {
   display: flex;
@@ -55,4 +56,88 @@
   font-weight: bold;
   display: flex;
   align-items: center;
+}
+
+.leave-btn {
+  margin-left: 1rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 8px 16px;
+  background: #3b3b3b;
+  border-radius: 8px;
+  font-weight: 500;
+  font-size: 17px;
+  color: #ffffff;
+  gap: 0.5rem;
+  cursor: pointer;
+}
+
+.exit-btn {
+  color: white;
+  background-color: #323332;
+  transform: scaleX(-1);
+  padding: 5px;
+  border-radius: 8px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+}
+
+.pop-content {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  font-family: Inter;
+  background: #323332;
+  border-radius: 8px;
+  padding: 24px;
+  color: white;
+}
+.pop-content p {
+  margin: 0;
+}
+.pop-head {
+  font-style: normal;
+  font-weight: 600;
+  font-size: 20px;
+  line-height: 28px;
+}
+.pop-text {
+  margin-top: 0.5rem;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 13px;
+  line-height: 16px;
+}
+.pop-overlay {
+  background-color: rgba(0, 0, 0, 0.7);
+  position: fixed;
+  inset: 0;
+}
+.cta-wrapper {
+  display: flex;
+  margin-top: 1.5rem;
+}
+.cta-wrapper button {
+  font-family: Inter;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 15px;
+  line-height: 24px;
+  line-height: 24px;
+  width: 150px;
+  cursor: pointer;
+}
+.leave-btn {
+  background: #ed4c5a;
+  border-radius: 8px;
+  margin-left: 0.5rem;
+}
+.cancel-btn {
+  color: white;
+  background: #444443;
+  border-radius: 8px;
 }
