+++ components/hms/footer/index.tsx
@@ -6,11 +6,13 @@ import {
   MicOffIcon,
   MicOnIcon,
   ShareScreenIcon,
-  RecordIcon
+  RecordIcon,
+  HangUpIcon
 } from '@100mslive/react-icons';
 import React from 'react';
 import s from './index.module.css';
 import Settings from './Settings';
+import * as Dialog from '@radix-ui/react-dialog';
 
 const Footer = () => {
   const role = useHMSStore(selectLocalPeerRole);
@@ -28,6 +30,13 @@ const Footer = () => {
       await actions.setScreenShareEnabled(true);
     } catch (error) {
       console.log('Error: ', error);
+    }
+  };
+  const leave = () => {
+    try {
+      actions.leave();
+    } catch (error) {
+      console.log(error);
     }
   };
   return (
@@ -67,6 +76,30 @@ const Footer = () => {
       {role?.name === 'backstage' || role?.name === 'stage' || role?.name === 'invitee' ? (
         <Settings />
       ) : null}
+      <Dialog.Root>
+        <Dialog.Overlay className={s['pop-overlay']} />
+        <Dialog.Trigger asChild>
+          <div className={s['btn-wrapper']}>
+            <button className={s['btn']} onClick={() => {}}>
+              <HangUpIcon />
+            </button>
+            <p className={s['btn-text']}>Leave</p>
+          </div>
+        </Dialog.Trigger>
+        <Dialog.Content className={s['pop-content']}>
+          <p className={s['pop-head']}>Leave the Stage</p>
+          <p className={s['pop-text']}>Are you sure you want to leave the stage?</p>
+          <div className={s['cta-wrapper']}>
+            <Dialog.Close asChild>
+              <button className={s['cancel-btn']}>Cancel</button>
+            </Dialog.Close>
+
+            <button className={s['leave-btn']} onClick={leave}>
+              Leave
+            </button>
+          </div>
+        </Dialog.Content>
+      </Dialog.Root>
     </div>
   );
 };
